+++ src/apic-extension/azext_apic_extension/aaz/latest/apic/__init__.py
@@ -11,7 +11,6 @@
 from .__cmd_group import *
 from ._create import *
 from ._delete import *
-from ._import import *
 from ._import_from_apim import *
 from ._list import *
 from ._show import *